Затащил минимальный базовый слушатель ошибок, также базовые ошибки.
