+++ app/pcheliki.monolit/src/Core/Enums/ErrorTypes.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Core\Enums;
+
+class ErrorTypes
+{
+    // По хорошему их бы к одному формату вывода ошибок привести, но пока так
+    public const PCHELIKI_VALIDATION_ERRORS = 0;
+
+    public const DEFAULT_VALIDATION_ERRORS = 1;
+}
