+++ app/pcheliki.monolit/src/Core/Exceptions/PchelikiNotFoundException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Core\Exceptions;
+
+// Ошибку выдаем когда такого ресурса не существует
+use Throwable;
+
+class PchelikiNotFoundException extends \Exception
+{
+    public function __construct(string $message = "", int $code = 0, ?Throwable $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+    }
+}
