+++ app/pcheliki.monolit/src/Core/Validator/IValidatorService.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Core\Validator;
+
+use Symfony\Component\Validator\Constraint;
+
+interface IValidatorService
+{
+    public function validateData(array $data, Constraint|array $constraint): ValidationResultTry;
+}
