+++ app/pcheliki.monolit/src/Modules/Auth/Dto/DtoRequestAuthCheckNumber.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Modules\Auth\Dto;
+
+final class DtoRequestAuthCheckNumber
+{
+    public function __construct(
+        public string $phone,
+        public bool $rememberMe,
+    ) {}
+}
