+++ app/pcheliki.monolit/src/Modules/Infrastructure/Base/BaseAuthController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Modules\Infrastructure\Base;
+
+class BaseAuthController extends BaseController
+{
+    //TODO Удалить если ничего нового не будет
+    public function __construct()
+    {
+        parent::__construct();
+    }
+
+}
