+++ app/pcheliki.monolit/src/Modules/Infrastructure/Base/BaseDefaultController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Modules\Infrastructure\Base;
+
+abstract class BaseDefaultController extends BaseController
+{
+    public function __construct()
+    {
+        parent::__construct();
+        // Тут скорее всего всякие мидлеваре будут
+    }
+
+}
