--- app/pcheliki.monolit/src/modules/Auth/Dto/DtoRequestAuthCheckNumber.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\modules\Auth\Dto;
-
-final class DtoRequestAuthCheckNumber
-{
-    public function __construct(
-        public string $phone,
-        public bool $rememberMe,
-    ) {}
-}
