--- app/pcheliki.monolit/src/modules/Default/Dto/DtoRequestDefault.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace App\modules\Default\Dto;
-final class DtoRequestDefault
-{
-
-    //Тут наши поля и маппинг их на дтошку array -> dto
-    public function __construct(
-        public string $default1,
-        public string $default2,
-        public string $default3,
-    ) {}
-
-
-}
