--- app/pcheliki.monolit/src/modules/Infrastructure/Base/BaseAuthController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\modules\Infrastructure\Base;
-
-class BaseAuthController extends BaseController
-{
-    //TODO Удалить если ничего нового не будет
-    public function __construct()
-    {
-        parent::__construct();
-    }
-
-}
