--- app/pcheliki.monolit/src/modules/Infrastructure/Validator/IValidatorService.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\modules\Infrastructure\Validator;
-
-use Symfony\Component\Validator\Constraint;
-
-interface IValidatorService
-{
-    public function validateData(array $data, Constraint|array $constraint): ValidationResultTry;
-}
