+++ app/pcheliki.monolit/templates/main/errors/access_denied.html.twig
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Доступ запрещен</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            background-color: #f5f5f5;
+            padding: 20px;
+            text-align: center;
+        }
+        h1 {
+            color: #333;
+        }
+        p {
+            color: #666;
+        }
+    </style>
+</head>
+<body>
+<h1>Доступ запрещен</h1>
+<p>Извините, у вас нет доступа к данному ресурсу</p>
+</body>
+</html>
