+++ app/pcheliki.monolit/templates/main/errors/not_found.html.twig
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Страница не найдена</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            background-color: #f5f5f5;
+            padding: 20px;
+            text-align: center;
+        }
+        h1 {
+            color: #333;
+        }
+        p {
+            color: #666;
+        }
+    </style>
+</head>
+<body>
+<h1>Страница не найдена</h1>
+<p>Извините, запрашиваемая страница не найдена.</p>
+</body>
+</html>
